perf(AreaTree): Optimize area searches to get O(log n) time
Ditto as last commit, and probably the last

Previous way of doing area searches not only was O(n), but
also iterated over a newly built list, rather than finding the area
during the in-order traversal.

The difference between this one and the previous commits is that this
one still retains a O(n) search method (although improved), since
searching for an area by its code is highly desirable.

This method is not logarithmic due to some constraints when building the
trees in Exercise1, as the areaName has to be used in the compareTo()
method of the Area class.