fix(ex1): Correctly add years/elements/items to areas
Previously, when adding a year to an element and that element to an
item, if the element already existed it would override the prevous
instance of element (due to the AVL's insert method), effectively
deleting the year previously stored inside such element.

This would cause exercise 3 to produce an incorrect answer, since it
would be impossible to find the latest year for which each area would
have records of a given item+element combination.